replace server file, include api endpoint handlers
